+++ Changelog.md
@@ -1,5 +1,6 @@
 ﻿v.1.1
 - Added the ability to vote to reroll a map.
+- Added Domination gamemode. Teams fight to control as many maps as possible before time runs out.
 - Visual pass to make UI elements more in line with UI changes from the latest game update.
 
 v.1.0.6